refactor(Introduction): Hoist 'acknowledged' state and add BackHandler
This commit refactors the `IntroductionScreen` to improve state management and navigation.

The `acknowledged` state has been hoisted from the `IntroductionScreen` composable and is now passed in as a `MutableState`. This allows parent composables, like `SetupCollectionFragment`, to control and observe this state.

A `BackHandler` has been added to the `IntroductionScreen`. When the "OK" button has been pressed (i.e., `acknowledged` is true), pressing the back button will now return the user to the initial view of the screen instead of closing the fragment.

To maintain backward compatibility within the codebase, an overload for `IntroductionScreen` has been added that internally creates and manages its own `acknowledged` state.