Refactor(MyAccount): Move back press handling into `MyAccountScreen`
This commit moves the `LaunchedEffect` responsible for managing the `onBackPressedCallback` from the `MyAccount` activity to the `MyAccountScreen` composable.

This change better co-locates the UI logic with the composable it affects, making the screen more self-contained and improving code organization. The `onBackPressedCallback` is now passed directly to `MyAccountScreen` as a parameter.