Refactor(Utils): Pre-allocate StringBuilder capacity in `ids2str`
This commit optimizes the `ids2str` methods by pre-allocating an estimated capacity for the `StringBuilder`.

By providing an initial size based on the expected length of the resulting ID string, this change reduces the number of internal re-allocations required by the `StringBuilder` as it grows, leading to a minor performance improvement.